Add assertions to code.
Specifically when user inputs a negative taskID or a taskID that
is does not exist (bigger than total number of tasks).

Having these assertions will improve code quality and will ensure
that command cases such as Mark, Unmark and Delete will work.
Previously, it was handled as an exception but now, it is
handled as an assertion.