Add a new class StringChecker to handle if task description is valid.
Previously, there were a lot of duplicated code and it in different
classes. I abstracted the code out and placed it in a static class
so that the code is cleaner and easier to read.

Following this, if there are any duplicated code in more classes
that requires checks, I can simply add them to the StringChecker
to practice good OOP.